Pedido - Microservicios

+++ Backend-Principal/src/main/java/com/bionaturista/domain/services/impl/PedidoServiceImpl.java
@@ -1,11 +1,12 @@
 package com.bionaturista.domain.services.impl;
 
+import com.bionaturista.application.dto.producto.ProductoDto;
+import com.bionaturista.application.dto.respuestas.Respuesta;
 import com.bionaturista.domain.Pattern.Publisher;
-import com.bionaturista.domain.Pattern.Subscriber;
 import com.bionaturista.domain.entities.*;
 import com.bionaturista.domain.repositories.*;
 import com.bionaturista.domain.services.PedidoService;
-import com.bionaturista.domain.services.UsuarioService;
+import com.bionaturista.domain.services.ProductoService;
 import com.bionaturista.validators.PedidoValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,21 +21,19 @@ public class PedidoServiceImpl implements PedidoService, Publisher {
     private InfoEnvioRepository infoEnvioRepository;
 
     @Autowired
-    private UsuarioServiceImpl service;
-
-    @Autowired
     private UsuarioRepository usuarioRepository;
-    //@Autowired
-    //private ProductoRepository productoRepository;
     @Autowired
     private EstadoPedidoRepository estadoPedidoRepository;
+    @Autowired
+    private UsuarioServiceImpl service;
+    private ProductoService productoService;
 
-    public PedidoServiceImpl(PedidoRepository pedidoRepository){
+    public PedidoServiceImpl(PedidoRepository pedidoRepository) throws InterruptedException{
         this.pedidoRepository=pedidoRepository;
     }
 
     @Override
-    public Pedido pagarPedido(Pedido pedido) {
+    public Pedido pagarPedido(Pedido pedido) throws InterruptedException {
         //Asigno el montoPago el costoEnvío para de ahí sumarle los productos jeje
         pedido.setMontoPago(pedido.getCostoEnvio());
         //Obtengo el Usuario mediante su idUsuario para obtener su carrito pex
@@ -47,9 +46,9 @@ public class PedidoServiceImpl implements PedidoService, Publisher {
         //Valida ☼
         //pedido.setProductosPedido(usuarioFinal.getCarritoCompras());
         PedidoValidator.validate(pedido);
-        //Vasea el carrito y se lo guarda al usuario (actualiza *en esencia*) ☼☼☼
-        //Set<Producto> nuevoCarrito = null;
-        //usuarioFinal.setCarritoCompras(nuevoCarrito);
+        //Vasea el carrito y se lo guarda al usuario (actualiza) ☼☼☼
+        Set<Producto> nuevoCarrito = null;
+        usuarioFinal.setCarritoCompras(nuevoCarrito);
 
         //Se debe agregar un metodo exits para verificar si la información de envio ya existe, pero me da pereza hacerlo
         InfoEnvio iv = this.infoEnvioRepository.save(pedido.getInfoEnvio());
@@ -58,29 +57,36 @@ public class PedidoServiceImpl implements PedidoService, Publisher {
 
         this.notificar(pedido.getUsuario(), "ha pagado su pedido");
 
-
         usuarioRepository.save(usuarioFinal);
         //Y guarda el pedido as it should :)
         return pedidoRepository.save(pedido);
     }
 
     @Override
-    public float obtenerMontoTotal(Usuario usuario, Pedido pedido) {
-        //Set<Producto> productos = usuario.getCarritoCompras();
-      //  float total=pedido.getMontoPago();
-        //for (Producto producto : productos){
+    public float obtenerMontoTotal(Usuario usuario, Pedido pedido) throws InterruptedException {
+        Set<Producto> productos = usuario.getCarritoCompras();
+        float total=pedido.getMontoPago();
+        for (Producto producto : productos){
             //Suma de precios
-           // total+=producto.getPrecioP();
+            total+=producto.getPrecioProducto();
             //Resta al stock de cada productazo
-            //producto.setStockP(producto.getStockP()-1);
-            //productoRepository.save(producto);
-        //}
-        //return total;
-        return 10;
+            producto.setStockProducto(producto.getStockProducto()-1);
+            //Hacemos el traspaso de Product a ProductDto para que se modifique bien lindo
+            ProductoDto productoDto = new ProductoDto();
+            productoDto.setIdProducto(producto.getIdProducto());
+            productoDto.setDescripcionProducto(producto.getDescripcionProducto());
+            productoDto.setImagenProducto(producto.getImagenProducto());
+            productoDto.setNombreProducto(producto.getNombreProducto());
+            productoDto.setCategoria(producto.getCategoria());
+            productoDto.setCompuesto(producto.getCompuesto());
+            //Guardamos el Producto con menos stockazo
+            productoService.modificarProducto(productoDto);
+        }
+        return total;
     }
 
     @Override
-    public float obtenerSubtotal(Pedido pedido) {
+    public float obtenerSubtotal(Pedido pedido) throws InterruptedException {
         float subtotal=0f;
         float montoCompleto=pedido.getMontoPago();
         float montoEnvio= pedido.getCostoEnvio();
@@ -89,55 +95,64 @@ public class PedidoServiceImpl implements PedidoService, Publisher {
     }
 
     @Override
-    public Pedido modificarPedido(Pedido pedido) {
+    public Pedido modificarPedido(Pedido pedido) throws InterruptedException {
         PedidoValidator.validate(pedido);
         this.notificar(pedido.getUsuario(), "ha modificado su pedido");
         return pedidoRepository.save(pedido);
     }
 
     @Override
-    public void eliminarPedido(Integer idPedido) {
-        Pedido pedido=pedidoRepository.findById(idPedido).orElse(new Pedido());
-        this.notificar(pedido.getUsuario(), "ha eliminado su pedido");
-        pedidoRepository.deleteById(idPedido);
+    public Respuesta eliminarPedido(Integer idPedido) throws InterruptedException {
+        Respuesta respuesta = new Respuesta();
+
+        if (pedidoRepository.existsById(idPedido)){
+            Pedido pedido=pedidoRepository.findById(idPedido).orElse(new Pedido());
+            this.notificar(pedido.getUsuario(), "ha eliminado su pedido");
+            pedidoRepository.deleteById(idPedido);
+            respuesta.setSatisfactorio(true);
+            respuesta.setCodigo("101");
+            respuesta.setMensaje("Pedido Eliminado.");
+            return respuesta;
+        }
+        else {
+            throw new InterruptedException("Error en la respuesta del Servicio Invocado.");
+        }
     }
 
     @Override
-    public Pedido obtenerPedidoPorIdPedido(Integer idPedido) {
-
+    public Pedido obtenerPedidoPorIdPedido(Integer idPedido) throws InterruptedException {
         return pedidoRepository.findById(idPedido).orElse(new Pedido());
-
     }
 
     @Override
-    public List<Pedido> listarPedido() {
+    public List<Pedido> listarPedido() throws InterruptedException {
         return pedidoRepository.findAll();
     }
 
     @Override
-    public Pedido modificarEstadoPedido(Integer idPedido) {
+    public Pedido modificarEstadoPedido(Integer idPedido) throws InterruptedException {
         return null;
     }
 
     @Override
-    public List<Pedido> listarPedidosPorIdUsuario(Usuario usuario) {
+    public List<Pedido> listarPedidosPorIdUsuario(Usuario usuario) throws InterruptedException {
         List<Pedido> usuarioId = pedidoRepository.listarPedidosPorIdUsuario(usuario);
         return usuarioId;
     }
 
     @Override
-    public void cancelarPedido(Integer idPedido) {
+    public void cancelarPedido(Integer idPedido) throws InterruptedException {
         Pedido pedido=pedidoRepository.findById(idPedido).orElse(new Pedido());
         EstadoPedido enviado=estadoPedidoRepository.pedidoEnviado();
         pedido.setEstadoPedido(enviado);
         this.notificar(pedido.getUsuario(), "ha cancelado su pedido");
     }
-
     @Override
     public void notificar(Usuario u2, String accion) {
         for (Usuario u:service.buscarUsuariosByRol("admin")) {
             service.actualizar(u, u2, accion);
         }
     }
+
 }
